+++ pudb/themes/mono.py
@@ -1,0 +1,7 @@
+palette_dict = {
+    "background": ("standout",),
+    "selectable": (),
+    "focused selectable": ("underline",),
+    "highlighted": ("bold",),
+    "hotkey": ("underline, standout",),
+}
